# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EpisodeStatus = typeof EpisodeStatus[keyof typeof EpisodeStatus];


export const EpisodeStatus = {
  draft: 'draft',
  complete: 'complete',
  review: 'review',
  approved: 'approved',
  scheduled: 'scheduled',
  published: 'published',
} as const;

export interface Episode {
  id: number;
  epNumber: number;
  status: EpisodeStatus;
  /** @nullable */
  dateBuilt?: string | null;
  postDate: string;
  season: string;
  aspectRatio?: string;
  duration: string;
  hookTitle: string;
  youtubeTitle: string;
  voScript: string;
  visualDirection: string;
  bgSound: string;
  thumbnailPrompt: string;
  citationCta: string;
  hashtags: string;
  /** @nullable */
  youtubeVideoId?: string | null;
  /** @nullable */
  scheduledPublishAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type EpisodeUpdateStatus = typeof EpisodeUpdateStatus[keyof typeof EpisodeUpdateStatus];


export const EpisodeUpdateStatus = {
  draft: 'draft',
  complete: 'complete',
  review: 'review',
  approved: 'approved',
  scheduled: 'scheduled',
  published: 'published',
} as const;

export interface EpisodeUpdate {
  status?: EpisodeUpdateStatus;
  youtubeTitle?: string;
  citationCta?: string;
  hashtags?: string;
  scheduledPublishAt?: string;
}

export type EpisodeStatsByStatus = {
  draft: number;
  complete: number;
  review: number;
  approved: number;
  scheduled: number;
  published: number;
};

export interface EpisodeStats {
  total: number;
  byStatus: EpisodeStatsByStatus;
  /** @nullable */
  nextPostDate?: string | null;
  publishedThisMonth: number;
}

export interface YouTubeAuthUrl {
  url: string;
}

export interface YouTubeStatus {
  connected: boolean;
  /** @nullable */
  channelName?: string | null;
  /** @nullable */
  channelId?: string | null;
}

export type PublishRequestPrivacyStatus = typeof PublishRequestPrivacyStatus[keyof typeof PublishRequestPrivacyStatus];


export const PublishRequestPrivacyStatus = {
  public: 'public',
  unlisted: 'unlisted',
  private: 'private',
} as const;

export interface PublishRequest {
  /**
     * ISO datetime to schedule. Null = publish immediately.
     * @nullable
     */
  scheduleAt?: string | null;
  privacyStatus: PublishRequestPrivacyStatus;
}

export interface PublishResult {
  success: boolean;
  /** @nullable */
  youtubeVideoId?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  message: string;
}

export type ListEpisodesParams = {
status?: ListEpisodesStatus;
season?: string;
};

export type ListEpisodesStatus = typeof ListEpisodesStatus[keyof typeof ListEpisodesStatus];


export const ListEpisodesStatus = {
  draft: 'draft',
  complete: 'complete',
  review: 'review',
  approved: 'approved',
  scheduled: 'scheduled',
  published: 'published',
} as const;