# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeStatus } from './episodeStatus';

export interface Episode {
  id: number;
  epNumber: number;
  status: EpisodeStatus;
  /** @nullable */
  dateBuilt?: string | null;
  postDate: string;
  season: string;
  aspectRatio?: string;
  duration: string;
  hookTitle: string;
  youtubeTitle: string;
  voScript: string;
  visualDirection: string;
  bgSound: string;
  thumbnailPrompt: string;
  citationCta: string;
  hashtags: string;
  /** @nullable */
  youtubeVideoId?: string | null;
  /** @nullable */
  scheduledPublishAt?: string | null;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}