# lib/api-zod/src/generated/types/episodeStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeStatsByStatus } from './episodeStatsByStatus';

export interface EpisodeStats {
  total: number;
  byStatus: EpisodeStatsByStatus;
  /** @nullable */
  nextPostDate?: string | null;
  publishedThisMonth: number;
}