# lib/api-zod/src/generated/types/episodeUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeUpdateStatus } from './episodeUpdateStatus';

export interface EpisodeUpdate {
  status?: EpisodeUpdateStatus;
  youtubeTitle?: string;
  citationCta?: string;
  hashtags?: string;
  scheduledPublishAt?: string;
}