# lib/api-zod/src/generated/types/publishRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PublishRequestPrivacyStatus } from './publishRequestPrivacyStatus';

export interface PublishRequest {
  /**
     * ISO datetime to schedule. Null = publish immediately.
     * @nullable
     */
  scheduleAt?: string | null;
  privacyStatus: PublishRequestPrivacyStatus;
}