# lib/api-zod/src/generated/types/publishResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface PublishResult {
  success: boolean;
  /** @nullable */
  youtubeVideoId?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  message: string;
}