# lib/api-zod/src/generated/types/youTubeStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * BioMinute Publishing Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface YouTubeStatus {
  connected: boolean;
  /** @nullable */
  channelName?: string | null;
  /** @nullable */
  channelId?: string | null;
}